posekit: split track_ui.py into recording helpers, theme strings, and the Gradio app
Pure relocation, no behaviour change:
- apis/track_recording.py: Session, the per-session Rerun recording (live
  stream tee + downloadable .rrd parts) and every overlay-logging helper;
  imports no gradio.
- track_ui_theme.py: DESCRIPTION / VIEWER_HEIGHT / FORCE_DARK_HEAD / APP_CSS.
- models/sam2_video.cached_predictor: the per-variant cached loader.
- track_ui.py keeps callbacks, build_demo, launch (834 -> 558 lines).
